+++ assets/js/one_chat.js
@@ -202,3 +202,9 @@ require('./chan_room')
 require('./messages')
 
 import * as cc from "./chat_channel"
+
+window.page_params = {
+  default_language: "en",
+  default_language_name: "English",
+  ucxchat: ucxchat,
+};