+++ lib/infinity_one_web/controllers/help_controller.ex
@@ -1,0 +1,9 @@
+defmodule InfinityOneWeb.HelpController do
+  use InfinityOneWeb, :controller
+
+  plug(:put_layout, false)
+
+  def index(conn, _params) do
+    render(conn, "index.html")
+  end
+end